refactor: change api clients
