+++ src/feature/api/api-client-anonymous.ts
@@ -1,14 +1,12 @@
 import { createApiBuilderFromCtpClient } from '@commercetools/platform-sdk';
-import type {
-  AuthMiddlewareOptions,
-  CorrelationIdMiddlewareOptions,
-} from '@commercetools/ts-client';
+import type { AuthMiddlewareOptions } from '@commercetools/ts-client';
 import { ClientBuilder } from '@commercetools/ts-client';
 
 import {
   authUrl,
   clientId,
   clientSecret,
+  correlationIdMiddlewareOptions,
   httpMiddlewareOptions,
   projectKey,
   scopes,
@@ -26,10 +24,6 @@ const AnonymousFlowApiClient = () => {
     scopes,
     tokenCache,
     httpClient: fetch,
-  };
-
-  const correlationIdMiddlewareOptions: CorrelationIdMiddlewareOptions = {
-    generate: (): string => crypto.randomUUID(),
   };
 
   const anonymousFlowClient = new ClientBuilder()
